requirements.txt
flask
flask-sqlalchemy
flask-restful
flask-marshmallow
marshmallow-sqlalchemy
flask_mysqldb